[step2] fix: 클래스 분리 (LottoProcessor, InputView)
