+++ src/main/java/lotto/LottoProcessor.java
@@ -2,32 +2,21 @@ package lotto;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Scanner;
 import java.util.stream.Collectors;
 
-// 로또의 입출력 담당
 public class LottoProcessor {
-    private static final Scanner SCANNER = new Scanner(System.in);
+    private final Lotto lotto;
 
-    public static int inputAmount() {
-        System.out.println("구입 금액을 입력해주세요.");
-        int purchaseAmount = SCANNER.nextInt();
-        SCANNER.nextLine();
-        System.out.println((purchaseAmount / Lotto.AMOUNT) + "개를 구매했습니다.");
-        return purchaseAmount;
+    public LottoProcessor(Lotto lotto) {
+        this.lotto = lotto;
     }
 
-    public static void printLotto(Lotto lotto) {
+    public void printLotto() {
         lotto.generateNumbers();
         lotto.printNumbers();
     }
 
-    public static String inputLastWinningNumbers() {
-        System.out.println("지난 주 당첨 번호를 입력해 주세요.");
-        return SCANNER.nextLine();
-    }
-
-    public static void getWinningStats(String inputNumbers, Lotto lotto) {
+    public void getWinningStats(String inputNumbers) {
         LottoWinningStats lottoWinningStats = new LottoWinningStats(getWinningNumbers(inputNumbers), lotto);
         lottoWinningStats.calculateWinningStats();
         lottoWinningStats.calculateWinningCounts();
@@ -36,7 +25,7 @@ public class LottoProcessor {
         System.out.println("총 수익률은 " + lottoWinningStats.getYield() + "입니다.");
     }
 
-    private static List<Integer> getWinningNumbers(String inputNumbers) {
+    private List<Integer> getWinningNumbers(String inputNumbers) {
         String inputNumbersTrim = inputNumbers.replaceAll(" ", "");
         String[] numbersString = inputNumbersTrim.split(",");
         List<Integer> winningNumbers = Arrays.stream(numbersString).map(Integer::parseInt).collect(Collectors.toList());
