+++ src/main/java/lotto/Main.java
@@ -2,12 +2,12 @@ package lotto;
 
 public class Main {
     public static void main(String[] args) {
-        int purchaseAmount = LottoProcessor.inputAmount();
+        int purchaseAmount = InputView.inputAmount();
         Lotto lotto = new Lotto(purchaseAmount);
+        LottoProcessor lottoProcessor = new LottoProcessor(lotto);
+        lottoProcessor.printLotto();
 
-        LottoProcessor.printLotto(lotto);
-
-        String winningNumbers = LottoProcessor.inputLastWinningNumbers();
-        LottoProcessor.getWinningStats(winningNumbers, lotto);
+        String winningNumbers = InputView.inputLastWinningNumbers();
+        lottoProcessor.getWinningStats(winningNumbers);
     }
 }
